// block-performance-1.0.0/src/library/web-client/get-file-temp/type.ts
export interface Input {
    url: string;
    headers?: { name: string; value: string }[];
    filename?: string;
}

export interface OutputUnexpectedResponse {
    status_code: number;
    status_message: string;
    headers: { name: string; value: string }[];
    temp_file_name: string;
}

export interface OutputOk {
    status_code: number;
    status_message: string;
    headers: { name: string; value: string }[];
    temp_file_name: string;
}

export interface OutputFail {}

export interface Callback<R> {
    unexpected_response: (output: OutputUnexpectedResponse) => R;
    ok: (output: OutputOk) => R;
    fail: (err: Error) => R;
}